+++ src/main/java/ch/x01/fuzzy/core/FuzzyEngineException.java
@@ -1,4 +1,4 @@
-package ch.x01.fuzzy.engine;
+package ch.x01.fuzzy.core;
 
 public class FuzzyEngineException extends RuntimeException {
