+++ src/test/java/ch/x01/fuzzy/api/FuzzyEngineTest.java
@@ -2,7 +2,7 @@ package ch.x01.fuzzy.api;
 
 import ch.x01.fuzzy.api.FuzzyEngine.InputVariable;
 import ch.x01.fuzzy.api.FuzzyEngine.OutputVariable;
-import ch.x01.fuzzy.engine.FuzzyEngineException;
+import ch.x01.fuzzy.core.FuzzyEngineException;
 import org.junit.Test;
 
 import static ch.x01.fuzzy.api.FuzzyModel.LinguisticVariable.lv;
