+++ src/test/java/ch/x01/fuzzy/parser/RuleParserTest.java
@@ -1,9 +1,9 @@
 package ch.x01.fuzzy.parser;
 
-import ch.x01.fuzzy.engine.FuzzyRule;
-import ch.x01.fuzzy.engine.FuzzyRuleStatus;
-import ch.x01.fuzzy.engine.LinguisticVariable;
-import ch.x01.fuzzy.engine.MembershipFunction;
+import ch.x01.fuzzy.core.FuzzyRule;
+import ch.x01.fuzzy.core.FuzzyRuleStatus;
+import ch.x01.fuzzy.core.LinguisticVariable;
+import ch.x01.fuzzy.core.MembershipFunction;
 import org.junit.Before;
 import org.junit.Test;
